repository: Use long-running filter process for converting LFS pointers
Wire up support for running gitaly-lfs-smudge as a long-running filter
process to speed up conversion of LFS pointers into their converted
objects. This should result in a signficant performance boost when
creating archives for large repositories with many LFS pointers and
avoid the "process-spawn storms" we saw in the past whenever such an
archive was requested.

Changelog: performance