cmd/gitaly-lfs-smudge: Move creation of the HTTP client
The GitLab HTTP client is currently created closely before performing
the request to retrieve the LFS pointers' smudged contents. This has
been fine until now given that we only ever requested a single LFS
pointer across the whole lifetime of the command. As we are about to
implement the long-running filter protocol though this will change so
that we may request an unbounded number of LFS pointers in a single
session.

Move out creation of the HTTP client so that it can be reused across
calls to `smudgeContents()`. This avoids some overhead like re-reading
contents of the GitLab secret from disk.

Note that this is essentially a change in behaviour: previously, if we
failed to create the HTTP client, then we'd have just silently returned
the blob's clean contents. This easily papers over misconfigurations
though and is thus not all that great. Now we do return errors in case
we failed to create the client, which means that any such misconfigured
instance may now error. This feels like a reasonable change though:
better to complain about errors than to silently swallow then.