hook: Add concurrency tracker
Add a concurrency tracker whose job it is to keep track of concurrent
calls based on some key type and key. This will be used for example, to
track the number of concurrent calls to pack objects based on user_id as
well as repository.

The LogConcurrency function returns a cleanup function,
which should get called when the caller finishes its process. When the
concurrent processes for a certain key is at 0, the element will get
removed from the map to save memory.

There are also two metrics that will give us increased insight to the
traffic patterns. The first metric gives us the total active callers
broken up by segment (repository, user_id, or even something else in the
future). This answers the question "how many unique ___"(user_id,
repository) have active pack object processes?

The second metric gives us a histogram of concurrent pack object
processes so we can answer the question "what's the p99 of concurrent
pack object processes for users? repositories?"