git: Change ReceiveHooksPayload to UserDetails
The ReceiveHooksPayload contains information like the UserID, Username,
Protocol that would also be useful for UploadPack. Rename this to
UserDetails to be more generic, so that we can share this in both places.

However, in order to keep upgrades compatible, keep the
ReceiveHooksPayload and the receive_hooks_payload json tag. During an
upgrade, the gitaly-hooks binary will be replaced before the gitaly
binary, so there will be a short period of time when an old gitaly will
be sending to a newer gitaly-hooks. For this reason, we want to allow
the newer gitaly-hooks to still recognize a receive_hooks_payload key.