linguist: Implement struct to store stats
We're about to introduce an implementation of getting the language
statistics in Go. For this we need a way to collect and store the
results in a cache. This cache will be used to incrementally calculate the
stats between commits.

This change introduces the linguist.languageStats struct which will deal
with all this.