FindChangedPaths: Remove --name-status and change parsing accordingly
Now that git diff-tree can potentially receive tree objects, in those
cases it will print the IDs of the trees as an extra first line of output
before the regular output. While this extra first line of output could
be suppressed for commits with --no-commit-id, there is no way to suppress
it for trees. So instead of trying to suppress this extra output, we allow
git diff-tree to print entire lines in its regular output by not passing
it --name-status anymore, and then we search for the beginning colon char
to find these entire lines of regular output, ignoring everything else
(the line with tree or commit IDs) before it. We then have to further
parse the entire lines of regular output using a regex though.