praefect: Add list-storages subcommand
It would be handy for Praefct admins to easily print out the virtual
storages with its associated storages without having to go into the
config. This change adds a new subcommand list-storages that either
prints out all virtual storages with their associated storages, or takes
a virtual storage as a flag and prints out storages for that virtual
storage.

Changelog: added