ssh: Convert helper to verify remote and local repo have same revision
The `cloneCommand.test()` function both executes the clone command and
then returns the `master` branch as well as tags of both local and
remote repositories. This is honestly a bit hard to read, and that isn't
helped by the fact that `test()` doesn't really convey anything.

Refactor the code to instead explicitly execute the command and then
manually compare revisions. Like this, we don't hide away what's
happening and thus ease the reading flow when going through tests.