ssh: Use table-driven tests instead of multiple indirections
Because SSHUploadPack can either be invoked with or without sidechannel
and with and without the pack-objects cache we need to re-run some tests
multiple times with all combinations. This is solved via separate tests
which call into the same function, but in a rather weird way where
`TestFn()` calls `testFn()`, which then again calls `testFn2()`. This is
mighty confusing.

Refactor the test to instead be table-driven to simplify it.