ssh: Verify we get expected errors instead of only asserting error codes
The tests verifying that request validation works as expected for
SSHUploadPack are currently only asserting the error code. This is
fragile and may easily lead to failures that we assume to be what we
expected, but which are actually caused by something different.

Refactor the test to instead verify the full error to tighten it.