ssh: Moderinze test exercising failure of Git in SSHUploadPack
Our testcase that exercises SSHUploadPack in the context of Git failures
is using quite dated code patterns. Modernize it to match our current
best practices.

While at it, refactor `testPostUploadPackFailedResponse()`: it's not at
all obvious what this function does just by looking at its name, and the
looping logic is a tad weird, too. This is both fixed by updating the
code style and renaming it to `recvUntilError()`, which should hopefully
be easier to understand.