gittest: Extend the protocol-detecting Git command factory
WWe have an intercepting Git command factory whose only purpose it is to
listen in on the `GIT_PROTOCOL` environment variable, but otherwise
execute the Git commands as we'd normally do. It's quite inflexible
though and only allows us to read the resulting protocol, without any
ability to also reset the current state again. This makes it hard to use
it in table-driven tests, where we may reuse the factory across multiple
test cases.

Refactor it into a proper type to be more extensible and add a new
`Reset()` function to allow resetting the currently recorded protocol
versions. This new functionality will be used in a subsequent commit.