ssh: Merge protocol tests into successful tests
The tests we have which exercise whether we can indeed use Git protocol
v2 as expected use the same setup as our other tests which exercise
successful invocations of SSHUploadPack, with the only exception being
that they don't use a protocol-detecting command factory.

Let's merge these tests, which also increases our test coverage for the
Git protocol to now also verify with sidechannels and the pack-objects
cache as well as without protocol v2.