ssh: Fix test setup not actually exercising clone failure
One of our tests that ought to verify whether cloning fails when we pass
an invalid repository only checked that an error was returned, but it
never checked what kind of error was returned. And sure enough, nowadays
we don't fail because of an invalid request, but because the test setup
is incomplete and missing the `gitaly-ssh` binary.

Fix the test setup to build this binary and verify that the error
matches what we expect so we don't regress again in the future.