Makefile: Consistently require `v` prefix for tool versions
While some of our tool versions require the `v` prefix of their tag to
be present, others don't. This is inconsistent and a tag confusing.

Unify this to always require the prefix. This also gives us more
flexibility given that it's now easily possible to e.g. override the
version with a commit ID.