gitlab: Add Features method that returns map of feature flags
Add a Features method on the client that calls the Rails api and contains
logic that turns the response from the API into a map of feature flags
and their associated boolean values.

We have a limitation with the current Rails api endpoint however. We
don't have a way to determine, the case when the feature gate is a
percentage of actors or time, or if its by project or group, whether or
not a feature is enabled. We can tell however, if a feature is
completely turned on or off.