ssh: Fix resource leak when spawning git-upload-pack(1) fails
In order to be able to monitor what git-upload-pack(1) is communicating
with the user we create a pipe so that we can intercept the data. And
while make sure to close the reading-side of the pipe in all cases, we
in fact don't close the writing side when spawning git-upload-pack(1)
fails.

Fix this potential resource leak by immediately deferring the close of
the writing side.