cgroups: Change metric type to counter
Change the cgroups metric type to counter rather than a guage, as the
cgroups metrics are really counters.

Also, remove the memory failed metric since rather than being a count of
number of failed processes, it functions more like a garbage collection.
It would be a misnomer to call it memory_failed.

Changelog: changed