cgroups: Adjust metric names
Change the metric names to be more accurate. The memory.failcnt tracks
how often the kernel has tried to reclaim memory. While this maps
closely to the number of killed processes, they are not exactly the same
thing.

The cgroups cpu usage metrics are a running total. Add "total" at the
end of the metric to indicate this.

Changelog: changed