Makefile: Adjust `FIPS_MODE` to match behaviour of other switches
The `FIPS_MODE` variable needs to be set to `1` in order to enable usage
of FIPS in our Makefile. This is different from all the other variables,
where we only verify that the variable is defined to a non-empty value.

Adjust the behaviour of this variable to match our other toggles. This
should be a backwards-compatible change: all existing sites that enable
FIPS modes are passing `FIPS_MODE=1`, which continues to behave the
same.