ci: Stop waiting for Postgres manually
GitLab CI will automatically wait for exposed ports of services to
become reachable, so it shouldn't be necessary to manually check whether
Postgres or PgBouncer services are up and running. Furthermore, this
check is problematic in the context of FIPS tests, where we don't have
the `psql` client available in our images.

Remove the check and let GitLab CI handle this for us.