operations: Remove check for specific error message
In UserMergeBranch we're verifying whether the error message returned by
`gitaly-git2go` contains a specific string to be able to discern the
case where the merge failed because of a merge conflict. This check is
not required anymore though given that we now always get a specific
`ConflictingFilesError{}` in that case.

Remove the check to clean up the logic.