repository: Fix test for voting in CreateRepositoryFromBundle on macOS
In `TestCreateFromBundle_transactional` we assert that the received
transactional votes match some static hashes so that we're sure that
they stay stable. This is breaking on macOS though: the second round of
votes asserts that hashing on-disk contents of the repository results in
a stable hash, but on macOS the on-disk contents are different because
the default Git configuration contains some additional keys not present
on other platforms.

Fix this issue by instead manually computing the expected hash by doing
the hashing ourselves first. Like this we continue to verify that we do
indeed hash the correct files in a deterministic order, but adapt to the
system's actual contents of those files.