command: Fix race with setting command names and context cancellation
When spawning commands, callers can afterwards call `SetMetricsCmd()`
and `SetMetricsSubCmd()` to have the command be properly represented in
Prometheus metrics via its command and subcommand. But because we spawn
a Goroutine that logs these metrics as soon as the context is done, this
creates a race between setting those names and context cancellation.

Fix this race by converting these functions into options that can be
passed to `New()` so that they're set up correctly at construction time.

Changelog: fixed