gittest: Fix `ExecOpts()` with explicit stdout writer
The `ExecOpts()` function accepts a configuration that allows the caller
to set various options like the standard streams. As it turns out though
setting `Stdout` in that configuration is broken and will cause us to
fail immediately. The root cause is that we call `exec.Cmd.Output()`,
which is not allowed when you set the command's `Stdout` variable.

Fix this by calling `Run()` instead when `Stdout` is passed by the
caller. Add tests to catch any future regressions.