smarthttp: Assert actual errors in validation tests for PostReceivePack
The tests which exercise our request validation for PostReceivePack only
assert error codes. This is quite fragile though as it could easily be
that the error code matches what we'd expect, while the actual error is
still caused by something different.

Refactor the test to instead verify that the actual error matches.