git: Always disable use of alternate refs for git-fetch(1)
While we know that git-fetch(1) is having issues when using alternate
refs, we currently only disable it when using `FetchInternal()`. Let's
globally disable usage of alternate refs so that we don't ever hit the
performance edge case when fetching into repositories which have an
object pool connected to them.

This fixes a performance regression in the FetchRemote RPC in case the
external configuration doesn't ask us to ignore alternate refs. In a
recent incident, latency of the RPC regressed from ~1s to ~90s per
invocation for forks of `gitlab-org/gitlab` when the external Git
configuration was removed.

Changelog: fixed