operations: Always enable structured errors in UserDeleteBranch
With 30d4fd427 (operations: Use structured errors in UserDeleteBranch,
2022-06-07), we have introduced structured errors in UserDeletBranch.
The intent is to both fix silent errors in case we failed to delete the
branch, and to fix replication in case no transactional votes were cast
in the error case.

The change has been rolled out on June 14th and is part of v15.1. So
let's remove the feature flag guarding it.

Changelog: fixed