diff: Convert `RawDiff()` test to not use worktree
The use of worktrees in our tests is discouraged in favor of using test
helpers like `git.WriteCommit()` to set up required test data. One of
our tests for `RawDiff()` is using a worktree though to verify that the
generated diff can in fact be reapplied to a repository to round-trip to
the same tree again, and git-apply(1) requires a worktree to apply
patches.

Convert the test by instead using `gitaly-git2go apply` to apply the
patch. Note that we're forced to use a different set of commits to do
this test though: the current difference contains changes to binary
files, and these cannot be handled by `gitaly-git2go`.