diff: Convert `RawPatch()` test to not use worktree
The use of worktrees in our tests is discouraged in favor of using test
helpers like `git.WriteCommit()` to set up required test data. One of
our tests for `RawPatch()` is using a worktree though to verify that the
generated diff can in fact be reapplied to a repository to round-trip to
the same tree again, and git-apply(1) requires a worktree to apply
patches.

Convert the test to not use a worktree anymore by instead using the
`gitaly-git2go apply` command to apply the patch, which doesn't require
a worktree.