repository: Simplify helper to list compressed file contents
In order to list compressed archive's contents we're currently first
writing the archive to disk and then pass it to a helper function that
executes a binary depending on what archive format is used. The first
step to write the contents to disk is handled by the caller, which is a
bit unwieldy.

Refactor the helper function to write the file to disk itself in order
to prepare for a second caller of it. Ideally, we'd not write the file
to disk at all but instead just use the `archive` Go package to
enumerate contents. But that refactoring is left for another day.