repository: Clarify `GetArchive()` path injection test
It seemingly used to be possible to inject arguments to `GetArchive()`
via the path parameter so that an adversary could overwrite arbitrary
locations on disk. The test we have that verifies that this cannot
happen is a tad complex and not really obvious with regards to what's
happening.

Refactor the test and add documentation to simplify it. This is done in
preparation for a change where we stop using a worktree.