repository: Set up secure test server via helper
One of our tests verifies that we can call `CreateFork()` with a TLS
enabled server. The setup is currently done manually by assembling all
the necessary prerequisites, but this is indeed not necessary: the
`testserver` package handles the setup of TLS-enabled servers just fine
when the `TLSAddress` is set in the Gitaly configuration.

Use `runRepositoryService()` with a TLS-enabled configuration instead to
reduce the boilerplate. This also allows us to run these tests with
Praefect enabled.