repository: Fix leaking command in `cloneFromURLCommand()` tests
The test to verify that `cloneFromURLCommand()` behaves as expected is
spawning a git-clone(1) command, but doesn't make sure that this command
terminates as expected. This has resulted in failed pipelines due to the
process leak checker detecting this escaped process.

Fix this flake by killing the process by cancelling the context and then
waiting for the process to return.